VCST-1916: Rollback CustomerModel version, update manifest
feat: added the ability to notify customers about items abandoned in the shopping cart. The general settings of the module allow you to configure the launch of the check, the settings at the store level allow you to enable the functionality and manage the time when the cart is considered to be abandoned.